requirements.txt
paho-mqtt==1.5.1
influxdb==5.3.0
influxdb-client==1.39.0
python-dotenv==1.0.0